type_refactaring
