+++ frontend/app/cheers/create/page.tsx
@@ -1,6 +1,7 @@
 // frontend/app/cheers/create/page.tsx
 import { cookies } from "next/headers";
-import { getCheerPresets, createCheer } from "@/lib/server/cheers";
+import { createCheer } from "@/lib/server/cheers";
+import { getCheerPresets } from "@/lib/server/cheerPresets";
 import CheerForm from "@/components/cheer/CheerForm";
 import { redirect } from "next/navigation";
 import type { CheerFormState } from "@/lib/types/cheer";
@@ -10,7 +11,7 @@ export default async function CreateCheerPage() {
   // JWT取得
   const cookieStore = await cookies();
   const token = cookieStore.get("__session")?.value ?? "";
-  console.log("token:", token); // ←ここを追加
+  console.log("token:", token); 
 
   // プリセット値をサーバーで取得
   const { cheerTypes, muscles, poses } = await getCheerPresets(token);
