+++ frontend/lib/server/cheers.ts
@@ -2,7 +2,6 @@
 // Next.js 15 サーバー専用fetchユーティリティ
 import "server-only";
 import { Cheer,CheerFormState } from "@/lib/types/cheer";
-import { CheerType, Muscle, Pose } from "@/lib/types/prests";
 
 // 一覧取得（SSR/Server Action用）
 export async function getCheers(token: string): Promise<Cheer[]> {
@@ -12,29 +11,6 @@ export async function getCheers(token: string): Promise<Cheer[]> {
   });
   if (!res.ok) throw new Error("cheers取得に失敗しました");
   return res.json();
-}
-
-// プリセット一括取得
-export async function getCheerPresets(token: string): Promise<{
-  cheerTypes: CheerType[];
-  muscles: Muscle[];
-  poses: Pose[];
-}> {
-  const [cheerTypes, muscles, poses] = await Promise.all([
-    fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/api/v1/cheer_types`, {
-      headers: { Authorization: `Bearer ${token}` },
-      cache: "no-store",
-    }).then(r => r.json()),
-    fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/api/v1/muscles`, {
-      headers: { Authorization: `Bearer ${token}` },
-      cache: "no-store",
-    }).then(r => r.json()),
-    fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/api/v1/poses`, {
-      headers: { Authorization: `Bearer ${token}` },
-      cache: "no-store",
-    }).then(r => r.json()),
-  ]);
-  return { cheerTypes, muscles, poses };
 }
 
 // 作成
